porting to real board

+++ includes/types.h
@@ -10,7 +10,7 @@ typedef signed char			int8_t;
 typedef signed short		int16_t;
 typedef signed int			int32_t;
 typedef signed long long	int64_t;
-typedef unsigned long long	uintptr_t;
+typedef unsigned int		uintptr_t;
 
 typedef unsigned long		size_t;
